update: modified root dir, 'app' -> 'nexus_agent'
close #17
