+++ nexus_agent/graph/nodes/base.py
