+++ nexus_agent/graph/nodes/report_assistant.py
@@ -1,35 +1,30 @@
 from langgraph.prebuilt import create_react_agent
-from langchain_community.tools.tavily_search import TavilySearchResults
 from langgraph.types import Command
+from langchain_community.tools.file_management.write import WriteFileTool
 from langchain_core.messages import HumanMessage
 
-from app.graph.nodes.base import Node
+from nexus_agent.graph.nodes.base import Node
 
 
-class NewsSearcherNode(Node):
+class ReportAssistantNode(Node):
     def __init__(self):
         super().__init__()
-        self.system_prompt = (
-            "You are a new search agent. search for US news from BBC Source only."
-            "Do nothing else"
-        )
         self.agent = None
 
     def _run(self, state: dict) -> dict:
         if self.agent is None:
-            tools = [TavilySearchResults(max_results=2)]
+            tools = [WriteFileTool()]
             llm = state["llm"]
             self.agent = create_react_agent(
                 llm,
                 tools,
-                prompt=self.system_prompt,
             )
         result = self.agent.invoke(state)
         return Command(
             update={
                 "messages": [
                     HumanMessage(
-                        content=result["messages"][-1].content, name="news_searcher"
+                        content=result["messages"][-1].content, name="report_assistant"
                     )
                 ]
             },
