+++ nexus_agent/graph/nodes/supervisor.py
@@ -3,7 +3,7 @@ from typing import TypedDict
 from langgraph.graph import END
 from langgraph.types import Command
 
-from app.graph.nodes.base import Node
+from nexus_agent.graph.nodes.base import Node
 
 
 class SupervisorNode(Node):
