+++ nexus_agent/utils/tools.py
